ref(user): move commands to core
