+++ src-user/Resources/config/commands.xml
@@ -5,30 +5,7 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="fos_user.command.activate_user" class="FOS\UserBundle\Command\ActivateUserCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:activate" />
-        </service>
-        <service id="fos_user.command.change_password" class="FOS\UserBundle\Command\ChangePasswordCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:change-password" />
-        </service>
-        <service id="fos_user.command.create_user" class="FOS\UserBundle\Command\CreateUserCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:create" />
-        </service>
-        <service id="fos_user.command.deactivate_user" class="FOS\UserBundle\Command\DeactivateUserCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:deactivate" />
-        </service>
-        <service id="fos_user.command.demote_user" class="FOS\UserBundle\Command\DemoteUserCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:demote" />
-        </service>
-        <service id="fos_user.command.promote_user" class="FOS\UserBundle\Command\PromoteUserCommand">
-            <argument type="service" id="fos_user.util.user_manipulator" />
-            <tag name="console.command" command="fos:user:promote" />
-        </service>
+
     </services>
 
 </container>
